apps/web/rag/requirements.txt
fastapi
python-multipart
httpx
google-generativeai
pydantic
pydantic-settings
uvicorn
python-dotenv
pillow  # for image processing 